RUE-410: share scheduler core
